Clean Up Complete
There is still one folder that I can't move or rename: Fancy3DStuff>DOTween. It is referenced in a Camera script so if we rewrite camera, we can probably delete it.

There are also unused files (images, textures, fonts, etc) in the resources folder but we can delete those all at once when (if) we retexture.